Compile time and runtime failures are different
If I treat them as equal then the sequence

  1. Success
  2. Fail to compile the module
  3. Success

leaves me with the `0 examples, 0 failures` message, which is not
exactly ok. This happens because on step 3 `main --rerun` is run
but there were no failed specs and therefore nothing to rerun.

The solution is to run `main --rerun` only if runtime failure was
detected at least once in the sequence of failures.